migrated to XPLDevices
