+++ lib/library.json
@@ -1,0 +1,20 @@
+{
+  "name": "XPLDevices",
+  "version": "1.0.0",
+  "description": "Easy-to-start toolbox to create flight simulation devices based on the XPLDirect driver by CuriosityWorkshop",
+  "keywords": "xplane, flighsim, arduino",
+  "authors": [
+    {
+      "name": "Martin Ruskowski",
+      "email": "martin@ruskowski.de",
+      "maintainer": true
+    }
+  ],
+  "license": "GPLV3",
+  "homepage": "",
+  "dependencies": {
+    "adafruit/Adafruit MCP23017 Arduino Library": "^2.3.0"
+  },
+  "frameworks": "*",
+  "platforms": "*"
+}
