+++ src/main.cpp
@@ -1,18 +1,8 @@
 #include <Arduino.h>
 #include <MemoryFree.h>
-#include <DigitalIn.h>
-#include <Button.h>
-#include <Encoder.h>
-#include <Switch.h>
-#include <LedShift.h>
-#include <AnalogIn.h>
-#include <Timer.h>
-#include <XPLDirect.h>
+#include <XPLDevices.h>
 
-#define DEBUG 1
-
-// Interface
-XPLDirect XP(&Serial);
+// #define DEBUG 1
 
 // MUX 0
 Encoder encNavInner(0, 1, 2, 0, eEnc4Pulse);
@@ -118,8 +108,7 @@ long int sw_fuel_lever2 = 2;
 // Setup
 void setup()
 {
-  Serial.begin(XPLDIRECT_BAUDRATE);
-  XP.begin("G1000 MFD");
+  XPsetup("G1000 MFD");
 
   // Setup Multiplexers
   DigitalIn.setMux(1, 0, 2, 3);
@@ -310,20 +299,6 @@ void setup()
   XP.registerDataRef(F("aerobask/eng/sw_fuel_lever2"), XPL_READWRITE, 100, 0, &sw_fuel_lever2);
 }
 
-void handleCommand(Button *btn)
-{
-  btn->handle();
-  int cmdPush = btn->getCommand();
-  if (btn->pressed())
-  {
-    XP.commandStart(cmdPush);
-  }
-  if (btn->released())
-  {
-    XP.commandEnd(cmdPush);
-  }
-}
-
 void handleCommand(Button *btn, bool input)
 {
   btn->handle(input);
@@ -338,44 +313,12 @@ void handleCommand(Button *btn, bool input)
   }
 }
 
-void handleCommand(Encoder *enc)
-{
-  enc->handle();
-  if (enc->up())
-  {
-    XP.commandTrigger(enc->getCommand(eEncCmdUp));
-  }
-  if (enc->down())
-  {
-    XP.commandTrigger(enc->getCommand(eEncCmdDown));
-  }
-  int cmdPush = enc->getCommand(eEncCmdPush);
-  if (cmdPush >= 0)
-  {
-    if (enc->pressed())
-    {
-      XP.commandStart(cmdPush);
-    }
-    if (enc->released())
-    {
-      XP.commandEnd(cmdPush);
-    }
-  }
-}
-
-void handleCommand(Switch *sw)
-{
-  if (sw->handle())
-  {
-    XP.commandTrigger(sw->getCommand());
-  }
-}
-
 // Main loop
 void loop()
 {
-  XP.xloop();
-  DigitalIn.handle();
+//   XP.xloop();
+//   DigitalIn.handle();
+  XPloop();
   leds.handle();
 
   handleCommand(&encNavInner);
